refactor egui_web (break up lib.rs)
